Added command line argument in order to recreate database
Did recreation of db only in dev mode

+++ Program.cs
@@ -16,42 +16,45 @@ namespace RazorMvc
         public static void Main(string[] args)
         {
             var host = CreateHostBuilder(args).Build();
-
-            CreateDbIfNotExists(host);
+            bool recreateDb = args.Contains("--recreateDb");
+            InitializeDb(host, recreateDb);
 
             host.Run();
         }
 
-        private static void CreateDbIfNotExists(IHost host)
+        public static IHostBuilder CreateHostBuilder(string[] args) =>
+           Host.CreateDefaultBuilder(args)
+               .ConfigureWebHostDefaults(webBuilder =>
+               {
+                   webBuilder.UseStartup<Startup>();
+               });
+
+        private static void InitializeDb(IHost host, bool recreateDb)
         {
             using (var scope = host.Services.CreateScope())
             {
                 var services = scope.ServiceProvider;
+                var logger = services.GetRequiredService<ILogger<Program>>();
+
                 try
                 {
                     var context = services.GetRequiredService<InternDbContext>();
                     var webHostEnvironment = services.GetRequiredService<IWebHostEnvironment>();
-                    if (webHostEnvironment.IsDevelopment())
+                    if (webHostEnvironment.IsDevelopment() && recreateDb)
                     {
+                        logger.LogDebug("User requested to recreate database.");
                         context.Database.EnsureDeleted();
                         context.Database.EnsureCreated();
+                        logger.LogWarning("The database was recreated.");
                     }
 
                     SeedData.Initialize(context);
                 }
                 catch (Exception ex)
                 {
-                    var logger = services.GetRequiredService<ILogger<Program>>();
                     logger.LogError(ex, "An error occurred creating the DB.");
                 }
             }
         }
-
-        public static IHostBuilder CreateHostBuilder(string[] args) =>
-            Host.CreateDefaultBuilder(args)
-                .ConfigureWebHostDefaults(webBuilder =>
-                {
-                    webBuilder.UseStartup<Startup>();
-                });
     }
 }
